old-tests/submission/Microsoft/semantic/figcaption.html

<!DOCTYPE html>
<html><head>
 <meta content="text/html; charset=utf-8" http-equiv="content-type">
 <title>verify FIGCAPTION is a HTMLElement</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-figcaption-element.html#the-figcaption-element" rel="help">
</head>
<body>
 <div id="log"></div>
 <figure>
  <figcaption id="figcaption">This is just an example of a FIGCAPTION in a FIGURE:</figcaption>
 </figure>

 <script type="text/javascript">
  test(function()
  {
     assert_array_equals( "[object HTMLElement]",
                          document.getElementById("figcaption").toString() )
  })
 </script>


</body></html>
